move over fetch batch types to separate file
