+++ src/utils/types/index.ts
@@ -1,1 +1,2 @@
 export type { chainType } from './chainType';
+export * from './fetchBatch';
